+++ src/trx/game/output/state.c
@@ -109,14 +109,14 @@ void Output_SetFogEnd(const int32_t dist)
 void Output_SetupBelowWater(const bool underwater)
 {
     m_IsWaterEffect = true;
-    m_IsWibbleEffect = g_TRVersion < 3 && !underwater;
+    m_IsWibbleEffect = !underwater;
     m_IsShadeEffect = true;
 }
 
 void Output_SetupAboveWater(const bool underwater)
 {
     m_IsWaterEffect = false;
-    m_IsWibbleEffect = g_TRVersion < 3 && underwater;
+    m_IsWibbleEffect = underwater;
     m_IsShadeEffect = underwater;
 }